move foundry-tracing to nx package
