--- foundry-tracing-foundations/tsup.config.ts
@@ -1,12 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: ['src/index.ts'],
-  format: ['esm', 'cjs'],
-  outDir: 'dist',
-  clean: true,
-  dts: true,
-  splitting: false,
-  sourcemap: true,
-  tsconfig: 'tsconfig.build.json',
-});
