+++ packages/foundry-tracing-foundations/.dockerignore
