+++ packages/foundry-tracing-foundations/.gitignore
