+++ packages/foundry-tracing-foundations/package.json
@@ -26,16 +26,9 @@
     "README.md",
     "LICENSE"
   ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "type": "module",
-  "module": "dist/index.js",
-  "exports": {
-    ".": {
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
+  "type": "commonjs",
+  "main": "./index.cjs",
+  "types": "./src/index.d.ts",
   "scripts": {
     "test": "jest",
     "test:watch": "jest --watch",
