+++ packages/foundry-tracing-foundations/src/computeModule.ts
@@ -1,12 +1,12 @@
-import { ComputeModule } from "@palantir/compute-module";
+import { ComputeModule } from '@palantir/compute-module';
 // Schema Definitions for compute module
 // IMPORTANT:  @sinclair/typebox is required!!!
 // https://github.com/palantir/typescript-compute-module?tab=readme-ov-file#schema-registration
-import { Type } from "@sinclair/typebox";
+import { Type } from '@sinclair/typebox';
 import dotenv from 'dotenv';
 
-import { collectTelemetryFetchWrapper } from '@tracing/Tracing';
-import { ComputeModuleType, ModuleConfig } from "@tracing/types";
+import { collectTelemetryFetchWrapper } from './Tracing';
+import { ComputeModuleType, ModuleConfig } from './types';
 
 dotenv.config();
 
@@ -26,7 +26,7 @@ function getModuleConfig(): ModuleConfig {
       return { isTest: true };
     default: // development
       return { isTest: false };
-  };
+  }
 }
 
 function createComputeModule(): ComputeModuleType {
@@ -43,7 +43,7 @@ function createComputeModule(): ComputeModuleType {
       register: function (operation: string, handler: Function) {
         this.listeners[operation] = { type: 'response', listener: handler };
         return this;
-      }
+      },
     };
     console.log('returning mock module');
     return mockModule;
@@ -54,13 +54,15 @@ function createComputeModule(): ComputeModuleType {
     sources: {},
     definitions: { Trace: Schemas.Trace },
   })
-    .register("Trace", async ({ inputJSON }) => {
+    .register('Trace', async ({ inputJSON }) => {
       const result = await collectTelemetryFetchWrapper(inputJSON);
       return result;
     })
-    .on("responsive", () => console.log("Foundry Tracing Foundations is ready"));
+    .on('responsive', () =>
+      console.log('Foundry Tracing Foundations is ready')
+    );
 
-  module.on("responsive", () => {
+  module.on('responsive', () => {
     console.log(`${process.env.LOG_PREFIX} Module is now responsive`);
   });
