+++ packages/foundry-tracing-foundations/src/process-env.d.ts
