+++ packages/foundry-tracing-foundations/src/types.ts
