+++ packages/foundry-tracing-foundations/src/utils/uuid.ts
