+++ packages/foundry-tracing-foundations/test/__fixtures__/childTelemetryPayload.ts
