+++ packages/foundry-tracing-foundations/test/__fixtures__/mockDecorators.ts
