+++ packages/foundry-tracing-foundations/test/__fixtures__/telemetryPayload.ts
