+++ tsconfig.base.json
@@ -16,7 +16,7 @@
     "baseUrl": ".",
     "paths": {
       "@codestrap/developer-foundations.foundry-tracing-foundations": [
-        "foundry-tracing-foundations"
+        "packages/foundry-tracing-foundations/src/index.ts"
       ],
       "@codestrap/developer-foundations-agents-vickie-bennie": [
         "packages/agents/vickie-bennie/src/index.ts"
